+++ src/classification_ConfusionMatrix.cpp
@@ -15,13 +15,16 @@ using namespace Rcpp;
 //' @usage
 //' cmatrix(
 //'   actual,
-//'   predicted
+//'   predicted,
+//'   w
 //' )
 //'
 //' @param actual A <[factor]>-vector of [length] \eqn{n}, and \eqn{k} levels.
 //' @param predicted A <[factor]>-vector of [length] \eqn{n}, and \eqn{k} levels.
+//' @param w A <[numeric]>--vector of [length] \eqn{n}. [NULL] by default. If passed it will return a weighted confusion matrix.
 //'
 //' @example man/examples/scr_confusionmatrix.R
+//' @example man/examples/scr_wconfusionmatrix.R
 //' @family classification
 //'
 //' @inherit specificity details
@@ -46,26 +49,37 @@ using namespace Rcpp;
 // [[Rcpp::export]]
 Rcpp::NumericMatrix cmatrix(
     const Rcpp::IntegerVector& actual,
-    const Rcpp::IntegerVector& predicted) {
+    const Rcpp::IntegerVector& predicted,
+    const Rcpp::Nullable<Rcpp::NumericVector>& w = R_NilValue) {
 
-  /*
-   * This function generates a confusion matrix
-   * by incrementing locations based on the actual
-   * vector. If it includes NA values it will
-   * crash, so this function should check for missing
-   * values in both vectors. This is for later
-   */
+    // 1) declare the output matrix
+    // compiler doesn't recognize it in if-statements
+    Rcpp::NumericMatrix output;
 
-  // 1) get levels of the for the
-  // naming of the matrix
-  const Rcpp::CharacterVector& levels = actual.attr("levels");
+    // 1.1) extract levels (classes)
+    // from the actual-vector and the length
+    Rcpp::CharacterVector levels = actual.attr("levels");
+    int k = levels.length() + 1;
 
-  Rcpp::NumericMatrix output = Rcpp::wrap(confmat(actual, predicted));
+    // 2) Determine the weights
+    // these are what determines 
+    // the content of the confusion matrix
+     if (w.isNull()) {
+        
+        output = Rcpp::wrap(confusionMatrix<Eigen::MatrixXi>(actual, predicted, k));
 
-  // 4) set the dimnames of
-  // the confusion matrix
-  output.attr("dimnames") = Rcpp::List::create(levels, levels);
-  output.attr("class")    = "cmatrix";
+    } else {
+        
+        output = Rcpp::wrap(confusionMatrix<Eigen::MatrixXd>(actual, predicted, k, w));
 
-  return output;
+    }
+
+    // 3) preare the matrix output
+    // by adding column and rownames
+    // and classes
+    Rcpp::rownames(output) = levels;
+    Rcpp::colnames(output) = levels;
+    output.attr("class") = "cmatrix";
+
+    return output;
 }